requirements.txt
# Mathmetics
numpy
matplotlib
scipy

# Simulation
pygame
PyOpenGL
PyOpenGL_accelerate